+++ lib/src/main/java/org/asamk/signal/manager/TrustLevel.java
@@ -1,6 +1,7 @@
 package org.asamk.signal.manager;
 
 import org.whispersystems.signalservice.api.messages.multidevice.VerifiedMessage;
+import org.whispersystems.signalservice.internal.storage.protos.ContactRecord;
 
 public enum TrustLevel {
     UNTRUSTED,
@@ -14,6 +15,20 @@ public enum TrustLevel {
             TrustLevel.cachedValues = TrustLevel.values();
         }
         return TrustLevel.cachedValues[i];
+    }
+
+    public static TrustLevel fromIdentityState(ContactRecord.IdentityState identityState) {
+        switch (identityState) {
+            case DEFAULT:
+                return TRUSTED_UNVERIFIED;
+            case UNVERIFIED:
+                return UNTRUSTED;
+            case VERIFIED:
+                return TRUSTED_VERIFIED;
+            case UNRECOGNIZED:
+                return null;
+        }
+        throw new RuntimeException("Unknown identity state: " + identityState);
     }
 
     public static TrustLevel fromVerifiedState(VerifiedMessage.VerifiedState verifiedState) {